Title: 계단 오르기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241105/계단 오르기/climbing-stairs.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 using namespace std;
-long long dp[1001] = {};
+// long long dp[1001] = {};
+int dp[1001] = {};
 int n;
 int main() {
     cin >> n;
